Rename from notFoundEndpoint to notFoundError
